fix: do not decode every URI when invoking `toAbsolutePath`
